pyproject.toml
[project]
name = "motion-backend"
version = "0.1.0"
description = "Motion basketball coaching platform — FastAPI backend"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0.36",
    "greenlet>=3.1",
    "alembic>=1.14",
    "psycopg[binary]>=3.2",
    "asyncpg>=0.30",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "anthropic>=0.40",
    "httpx>=0.27",
    "python-jose[cryptography]>=3.3",
    "slowapi>=0.1.9",
    "structlog>=24.4",
    "pypdfium2>=5.7",
    "pyyaml>=6.0",
    "Pillow>=10,<13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "testcontainers[postgres]>=4.8",
    "ruff>=0.7",
    "mypy>=1.13",
    "types-python-jose",
    "types-PyYAML",
]

[project.scripts]
motion-api = "motion.main:run"
wiki-count-pages = "motion.wiki_ops.cli:count_pages"
wiki-check-nba-terms = "motion.wiki_ops.cli:check_nba_terms"
wiki-lint = "motion.wiki_ops.cli:lint_wiki"
wiki-detect-page-offsets = "motion.wiki_ops.cli:detect_page_offsets"
wiki-ingest = "motion.wiki_ops.cli:ingest"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/motion"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "S", "A", "C4", "PT", "SIM", "RUF"]
ignore = ["S101"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "PT"]
"alembic/**" = ["E501", "N806"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
# pypdfium2 does not currently ship PEP-561 type stubs; suppress
# "missing stubs" errors only for that single module so strict mode
# stays intact everywhere else.
module = "pypdfium2"
ignore_missing_imports = true

[[tool.mypy.overrides]]
# PIL/Pillow ships inline stubs since v10; suppress import errors on
# older mypy builds that haven't picked them up yet.
module = "PIL.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20260408",
]
